Updating MarkdownParse.java
Updated MarkdownParse.java to accommodate for the second way of typing in a link ([1] http://ucs.edu)

Read in the next character after the closed bracket and compared it to '(' or ':'.

If the next character is '(', print the substring inside the parenthesis

If the next character is ':', find index of . and +3 for the edu/com/org/etc. and print the substring from index of : to index of . + 3